fixed STL types export warnings, added checks to Deyo.h and Entrypoint.h

+++ Deyo/src/Platform/Windows/WinWindow.cpp
@@ -132,7 +132,7 @@ namespace Deyo
 		{
 			WindowData& winData = *(WindowData*)glfwGetWindowUserPointer(window);
 			
-			MouseScrollEvent evt(offsetX, offsetY);
+			MouseScrollEvent evt((float)offsetX, (float)offsetY);
 			winData.EventCallback(evt);
 		});
